Move path converting logic to base class since all detectors use it
