+++ plugin/src/main/groovy/com/btkelly/gnag/reporters/CheckstyleViolationDetector.groovy
@@ -17,7 +17,6 @@ package com.btkelly.gnag.reporters
 
 import com.btkelly.gnag.extensions.ReporterExtension
 import com.btkelly.gnag.models.Violation
-import com.btkelly.gnag.utils.DefaultRelativeFilePathComputer
 import com.puppycrawl.tools.checkstyle.ant.CheckstyleAntTask
 import groovy.util.slurpersupport.GPathResult
 import org.gradle.api.Project
@@ -26,7 +25,7 @@ import static com.btkelly.gnag.utils.StringUtils.sanitize
 /**
  * Created by bobbake4 on 4/1/16.
  */
-class CheckstyleViolationDetector extends BaseExecutedViolationDetector implements DefaultRelativeFilePathComputer {
+class CheckstyleViolationDetector extends BaseExecutedViolationDetector {
 
     CheckstyleViolationDetector(ReporterExtension reporterExtension, Project project) {
         super(reporterExtension, project)
@@ -78,7 +77,7 @@ class CheckstyleViolationDetector extends BaseExecutedViolationDetector implemen
                         sanitize((String) name()),
                         sanitize((String) violation.@message.text()),
                         null,
-                        computeFilePathRelativeToProjectRoot(sanitize((String) file.@name.text()), project),
+                        computeFilePathRelativeToProjectRoot(sanitize((String) file.@name.text())),
                         lineNumber))
             }
         }
